+++ lib/src/components/functions/set_value.dart
@@ -2,8 +2,8 @@ import 'package:json_dynamic_widget/json_dynamic_widget.dart';
 import 'package:json_dynamic_widget/src/components/types.dart';
 
 /// Function body for the built in `set_value` function. The `set_value`
-/// function accepts a key and a value and then calls the [setValue] function
-/// with those values.
+/// function accepts a key and a value and then returns a [VoidCallback] that
+/// calls the [setValue] function with those values when executed.
 ///
 /// The `set_value` function takes two values in the `args` array:
 ///  1. [String] -- the key to pass to [setValue].