+++ test/builders/json_wrap_builder_test.dart
@@ -1,0 +1,21 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:json_dynamic_widget/json_dynamic_widget.dart';
+
+void main() {
+  test('type', () {
+    const type = JsonWrapBuilder.type;
+
+    expect(type, 'wrap');
+    expect(
+      JsonWidgetRegistry.instance.getWidgetBuilder(type) is Function,
+      true,
+    );
+
+    expect(
+      JsonWidgetRegistry.instance.getWidgetBuilder(type)(
+        {},
+      )! is JsonWrapBuilder,
+      true,
+    );
+  });
+}